requirements.txt
streamlit
torch
transformers==4.39.3
accelerate==0.28.0
sentence-transformers==2.6.1
peft==0.10.0
pymupdf
faiss-cpu
